Add files via upload
This project is design by using HTML, CSS, JAVASCRIPT.In this project a simple counter is there which count the number from zero. There are three option given viz-Increment which increment the counter by the 1,Decrement which decrease the counter by 1,Neutral which make the Incremented or decremented value of counter to zero.